kafka: cleanup librdkafka utils
