+++ contrib/kafka/filters/network/source/mesh/librdkafka_utils.h
@@ -1,0 +1,54 @@
+#pragma once
+
+#include <map>
+#include <memory>
+#include <string>
+#include <utility>
+
+#include "envoy/common/pure.h"
+
+#include "absl/strings/string_view.h"
+#include "librdkafka/rdkafkacpp.h"
+
+namespace Envoy {
+namespace Extensions {
+namespace NetworkFilters {
+namespace Kafka {
+namespace Mesh {
+
+/**
+ * Helper class responsible for creating librdkafka entities, so we can have mocks in tests.
+ */
+class LibRdKafkaUtils {
+public:
+  virtual ~LibRdKafkaUtils() = default;
+
+  virtual RdKafka::Conf::ConfResult setConfProperty(RdKafka::Conf& conf, const std::string& name,
+                                                    const std::string& value,
+                                                    std::string& errstr) const PURE;
+
+  virtual RdKafka::Conf::ConfResult setConfDeliveryCallback(RdKafka::Conf& conf,
+                                                            RdKafka::DeliveryReportCb* dr_cb,
+                                                            std::string& errstr) const PURE;
+
+  virtual std::unique_ptr<RdKafka::Producer> createProducer(RdKafka::Conf* conf,
+                                                            std::string& errstr) const PURE;
+
+  virtual std::unique_ptr<RdKafka::KafkaConsumer> createConsumer(RdKafka::Conf* conf,
+                                                                 std::string& errstr) const PURE;
+
+  // Returned type is a raw pointer, as librdkafka does the deletion on successful produce call.
+  virtual RdKafka::Headers* convertHeaders(
+      const std::vector<std::pair<absl::string_view, absl::string_view>>& headers) const PURE;
+
+  // In case of produce failures, we need to dispose of headers manually.
+  virtual void deleteHeaders(RdKafka::Headers* librdkafka_headers) const PURE;
+};
+
+using RawKafkaConfig = std::map<std::string, std::string>;
+
+} // namespace Mesh
+} // namespace Kafka
+} // namespace NetworkFilters
+} // namespace Extensions
+} // namespace Envoy
