+++ contrib/kafka/filters/network/source/mesh/upstream_kafka_client_impl.h
@@ -4,42 +4,14 @@
 
 #include "envoy/event/dispatcher.h"
 
+#include "contrib/kafka/filters/network/source/mesh/librdkafka_utils.h"
 #include "contrib/kafka/filters/network/source/mesh/upstream_kafka_client.h"
-#include "librdkafka/rdkafkacpp.h"
 
 namespace Envoy {
 namespace Extensions {
 namespace NetworkFilters {
 namespace Kafka {
 namespace Mesh {
-
-/**
- * Helper class responsible for creating librdkafka entities, so we can have mocks in tests.
- */
-class LibRdKafkaUtils {
-public:
-  virtual ~LibRdKafkaUtils() = default;
-
-  virtual RdKafka::Conf::ConfResult setConfProperty(RdKafka::Conf& conf, const std::string& name,
-                                                    const std::string& value,
-                                                    std::string& errstr) const PURE;
-
-  virtual RdKafka::Conf::ConfResult setConfDeliveryCallback(RdKafka::Conf& conf,
-                                                            RdKafka::DeliveryReportCb* dr_cb,
-                                                            std::string& errstr) const PURE;
-
-  virtual std::unique_ptr<RdKafka::Producer> createProducer(RdKafka::Conf* conf,
-                                                            std::string& errstr) const PURE;
-
-  // Returned type is a raw pointer, as librdkafka does the deletion on successful produce call.
-  virtual RdKafka::Headers* convertHeaders(
-      const std::vector<std::pair<absl::string_view, absl::string_view>>& headers) const PURE;
-
-  // In case of produce failures, we need to dispose of headers manually.
-  virtual void deleteHeaders(RdKafka::Headers* librdkafka_headers) const PURE;
-};
-
-using RawKafkaProducerConfig = std::map<std::string, std::string>;
 
 /**
  * Combines the librdkafka producer and its dedicated monitoring thread.
@@ -53,11 +25,11 @@ class RichKafkaProducer : public KafkaProducer,
 public:
   // Main constructor.
   RichKafkaProducer(Event::Dispatcher& dispatcher, Thread::ThreadFactory& thread_factory,
-                    const RawKafkaProducerConfig& configuration);
+                    const RawKafkaConfig& configuration);
 
   // Visible for testing (allows injection of LibRdKafkaUtils).
   RichKafkaProducer(Event::Dispatcher& dispatcher, Thread::ThreadFactory& thread_factory,
-                    const RawKafkaProducerConfig& configuration, const LibRdKafkaUtils& utils);
+                    const RawKafkaConfig& configuration, const LibRdKafkaUtils& utils);
 
   // More complex than usual.
   // Marks that monitoring thread should finish and waits for it to join.
